限制登录人数为1，jwttoken过期时间为十分钟

+++ src/main/java/com/project/config/filter/RestfulFilter.java
@@ -11,7 +11,6 @@ import com.project.service.TokenManager;
 import com.project.util.JwtUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.shiro.authc.UsernamePasswordToken;
-import org.apache.shiro.subject.Subject;
 import org.apache.shiro.web.filter.authc.UserFilter;
 import org.apache.shiro.web.util.WebUtils;
 import org.slf4j.Logger;
@@ -41,25 +40,32 @@ public class RestfulFilter extends UserFilter {
         String loginToken = getToken(request);
 
         Optional<String> userName = JwtUtil.getUserName(loginToken);
+        //比较refreshToken的时间戳
+        if(userName.isPresent()) {
+            // 获取RefreshToken时间戳,及AccessToken中的时间戳
+            String refreshTokenCacheKey = RedisKey.REFRESH_TOKEN_PREFIX + userName.get();
+            String buildMillis = JwtUtil.getClaims(loginToken, RedisKey.CURRENT_TIME_MILLIS).orElse("");
+            String reFreshMillis = (String) redisManager.get(refreshTokenCacheKey);
+            //验证token使用者的唯一性，只能允许一个人使用
+            if (buildMillis.equals(reFreshMillis)) {
+                return true;
+            }else{
+                logger.warn("token已失效");
+                return false;
+            }
+        }
         //jwtToken失效情况,查看redis是否还存有用户信息
-        if(!userName.isPresent()) {
+        else{
             Optional<UsernamePasswordToken> user = Optional.ofNullable(tokenManager.getToken(loginToken));
             //用户信息还在说明 jwttoken过期了
             if(user.isPresent()) {
-//                if (userName.equals(token.get().getUsername())) {
-//                    long buildTime = Long.parseLong(JwtUtil.getClaims(loginToken, RedisKey.CURRENT_TIME_MILLIS).orElse(""));
-//                    long diffTime = (System.currentTimeMillis() - buildTime) / 1000;
-//                    if (diffTime > RedisKey.FIVE_MINIUTE) {
-                        refreshToken(user.get(),loginToken, response);
-//                    }
-                    return true;
-//                }
+                refreshToken(user.get(),loginToken, response);
+                return true;
+            }else{
+                logger.warn("token验证失败");
+                return false;
             }
-        }else{
-            return true;
         }
-        logger.warn("token验证失败");
-        return false;
     }
 
     /**
@@ -82,28 +88,29 @@ public class RestfulFilter extends UserFilter {
      */
     private boolean refreshToken(UsernamePasswordToken user,String oldToken,ServletResponse response) {
         // 获取当前Token的帐号信息
-//        String userName = JwtUtil.getUserName(token);
-//        String refreshTokenCacheKey = RedisKey.REFRESH_TOKEN_PREFIX + userName;
-        // 判断Redis中RefreshToken是否存在
-//        if (redisManager.hasKey(refreshTokenCacheKey)) {
-//            // 获取RefreshToken时间戳,及AccessToken中的时间戳
-//            // 相比如果一致，进行AccessToken刷新
-//            String currentTimeMillisRedis = (String) redisManager.get(refreshTokenCacheKey);
-//            if (String.valueOf(buildMillis).equals(currentTimeMillisRedis)) {
+        String userName = user.getUsername();
+        String refreshTokenCacheKey = RedisKey.REFRESH_TOKEN_PREFIX + userName;
+//         判断Redis中RefreshToken是否存在
+        if (redisManager.hasKey(refreshTokenCacheKey)) {
+            // 获取RefreshToken时间戳,及AccessToken中的时间戳
+            // 相比如果一致，进行AccessToken刷新
+
+//            String buildMillis = JwtUtil.getClaims(oldToken, RedisKey.CURRENT_TIME_MILLIS).orElse("");
+//            if (buildMillis.equals(currentTimeMillisRedis)) {
                 // 设置RefreshToken中的时间戳为当前最新时间戳
-//                String currentTimeMillis = String.valueOf(System.currentTimeMillis());
                 tokenManager.deleteToken(oldToken);
-                tokenManager.saveToken(user);
-                // 刷新AccessToken，为当前最新时间戳
-                String token = JwtUtil.buildJWT(user.getUsername(), RedisKey.TEN_MINIUTE);
+                Token token = tokenManager.saveToken(user);
                 // 设置响应的Header头新Token
                 HttpServletResponse httpServletResponse = (HttpServletResponse) response;
-                httpServletResponse.setHeader("Authentication", token);
+                httpServletResponse.setHeader("Authentication", token.getToken());
 //                httpServletResponse.setHeader("Access-Control-Expose-Headers", SecurityConsts.REQUEST_AUTH_HEADER);
-
+                return true;
+//            }else{
+//                return false;
 //            }
-//        }
-        return true;
+        }else{
+            return false;
+        }
     }
 
     /**
